backend/cloud-functions/IssueSearchAlgo/src/requirements.txt
functions-framework
pymongo==4.7.2
python-dotenv
